+++ releasenotes.md
@@ -10,3 +10,4 @@ Whats New:
 - Username is now removed, use your email address to login instead
 - User list page removed, we only allow a single user
 - Allow a user to register a PV Online account during a first time installation wizard
+- Add PV Online connectivity status (Online/Offline)